requirements.txt
python-telegram-bot
google-generativeai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
Flask
python-dotenv